# src/main/resources/swagger.yaml
info:
  title: "SPS Friend Management"
  version: "1.0"
openapi: 3.0.1
paths:
  /friends/common:
    get:
      operationId: getCommonFriends
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Friends'
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessage'
  /friends/subscribe:
    post:
      operationId: addSubscription
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestModel'
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessage'
  /friends/register:
    post:
      operationId: addFriends
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Friends'
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessage'
  /friends/add:
    post:
      operationId: addFriendships
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Friends'
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessage'
  /friends/block:
    put:
      operationId: addBlock
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestModel'
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessage'
  /friends/notify:
    get:
      operationId: addBlock_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotifyModel'
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessage'
  /friends/list:
    get:
      operationId: getFriendsByEmail
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Email'
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessage'
components:
  schemas:
    Message:
      type: object
      properties:
        friends:
          type: array
          items:
            type: string
        count:
          type: integer
          format: int32
        message:
          type: string
        recipients:
          type: array
          items:
            type: string
        success:
          type: boolean
    Friends:
      type: object
      properties:
        friends:
          type: array
          items:
            type: string
    RequestModel:
      type: object
      properties:
        requester:
          type: string
        target:
          type: string
    NotifyModel:
      type: object
      properties:
        sender:
          type: string
        text:
          type: string
    Email:
      type: object
      properties:
        email:
          type: string